add codex theme add obfuscated results
