+++ package.json
@@ -61,7 +61,7 @@
     "jest-junit": "^12.0.0",
     "jest-preset-angular": "^8.3.1",
     "lint-staged": "^10.5.0",
-    "lodash": "^4.17.20",
+    "lodash": "^4.17.21",
     "prettier": "^2.1.2",
     "pretty-quick": "^3.1.0",
     "protractor": "~7.0.0",
