+++ src/app/layout/components/app-layout/app-layout.component.html
@@ -1,3 +1,4 @@
+<num-header role="banner"></num-header>
 <mat-sidenav-container autosize fullscreen class="sidenav-container fix-overlay">
   <mat-sidenav
     #drawer
@@ -17,7 +18,43 @@
 
   <mat-sidenav-content fxLayout="column" class="fix-overlay">
     <!-- Page Content -->
-    <num-header role="banner"></num-header>
+
+    <section role="heading" fxLayout="row" fxLayoutAlign="center start" *ngIf="currentMainNavItem">
+      <div
+        class="titlebar num-margin-b-60"
+        role="presentation"
+        fxLayout="column"
+        fxLayoutAlign="center start"
+        fxFlex="95%"
+        fxFlex.gt-md="90%"
+        fxFlex.gt-lg="1200px"
+      >
+        <h1>
+          <fa-icon
+            class="num-fc--a"
+            [icon]="currentMainNavItem.icon"
+            size="sm"
+            [fixedWidth]="true"
+          ></fa-icon>
+          {{ currentMainNavItem.translationKey | translate }}
+        </h1>
+
+        <nav mat-tab-nav-bar [backgroundColor]="'#fff'" color="accent" *ngIf="currentTabNav">
+          <a
+            class="no-focus"
+            mat-tab-link
+            *ngFor="let tab of currentTabNav"
+            [routerLink]="tab.routeTo"
+            routerLinkActive
+            #rla="routerLinkActive"
+            [routerLinkActiveOptions]="{ exact: true }"
+            [active]="tab.id === currentTabNavSelected"
+          >
+            {{ tab.translationKey | translate }}
+          </a>
+        </nav>
+      </div>
+    </section>
 
     <div class="safari-fix" fxLayout="row" fxLayoutAlign="center start" role="presentation">
       <main
