+++ src/app/layout/components/footer/footer.component.scss
@@ -2,12 +2,22 @@
   font-size: 14px;
   height: 90px;
 }
-
+.mat-toolbar-row {
+  height: inherit;
+}
 .version-row {
   height: 32px;
   font-size: 12px;
 }
 
 .version {
-  padding: 0 16px 0 0;
+  text-align: center;
+}
+.bmbf-logo {
+  padding-top: 10px;
+  max-width: 100px;
+}
+.copyright {
+  text-align: right;
+  margin-right: 14px;
 }
