+++ src/app/layout/components/header/header.component.ts
@@ -1,8 +1,4 @@
-import { Component, OnDestroy, OnInit } from '@angular/core'
-import { ActivationEnd, Router, RouterEvent } from '@angular/router'
-import { Subscription } from 'rxjs'
-import INavItem from '../../models/nav-item.interface'
-import { mainNavItems } from '../../../core/constants/navigation'
+import { Component, OnInit } from '@angular/core'
 import { OAuthService, UserInfo } from 'angular-oauth2-oidc'
 
 @Component({
@@ -10,23 +6,12 @@ import { OAuthService, UserInfo } from 'angular-oauth2-oidc'
   templateUrl: './header.component.html',
   styleUrls: ['./header.component.scss'],
 })
-export class HeaderComponent implements OnInit, OnDestroy {
-  private subscriptions = new Subscription()
-
-  mainNavItems = mainNavItems
-  currentNavId: string
-  currentMainNavItem: INavItem
-  currentTabNav: INavItem[] = null
-  currentTabNavSelected: string
-
+export class HeaderComponent implements OnInit {
   profile: UserInfo
 
-  constructor(private router: Router, private oauthService: OAuthService) {}
+  constructor(private oauthService: OAuthService) {}
 
   ngOnInit(): void {
-    this.subscriptions.add(
-      this.router.events.subscribe((event) => this.handleRouterEvent(event as RouterEvent))
-    )
     this.initProfile()
   }
 
@@ -35,27 +20,5 @@ export class HeaderComponent implements OnInit, OnDestroy {
     if (isLoggedIn) {
       this.profile = await this.oauthService.loadUserProfile()
     }
-  }
-
-  ngOnDestroy(): void {
-    this.subscriptions.unsubscribe()
-  }
-
-  handleRouterEvent(routerEvent: RouterEvent): void {
-    if (routerEvent instanceof ActivationEnd) {
-      this.currentTabNavSelected = routerEvent.snapshot.firstChild?.data?.tabNavId
-
-      const navId = routerEvent.snapshot.data.navId
-      if (navId !== this.currentNavId) {
-        this.currentNavId = navId
-        this.setHeader()
-      }
-    }
-  }
-
-  setHeader(): void {
-    const navItem = this.mainNavItems.find((item) => item.routeTo === this.currentNavId)
-    this.currentMainNavItem = navItem
-    this.currentTabNav = navItem?.tabNav
   }
 }
