+++ src/app/layout/font-awesome-icons.ts
@@ -36,7 +36,8 @@ import { faGreaterThanEqual } from '@fortawesome/free-solid-svg-icons'
 import { faNotEqual } from '@fortawesome/free-solid-svg-icons'
 import { faSpinner } from '@fortawesome/free-solid-svg-icons'
 import { faWrench } from '@fortawesome/free-solid-svg-icons'
-
+import { faAngleDoubleLeft } from '@fortawesome/free-solid-svg-icons'
+import { faAngleDoubleRight } from '@fortawesome/free-solid-svg-icons'
 export const FONT_AWESOME_ICONS = [
   faTh,
   faMicroscope,
@@ -76,4 +77,6 @@ export const FONT_AWESOME_ICONS = [
   faNotEqual,
   faSpinner,
   faWrench,
+  faAngleDoubleLeft,
+  faAngleDoubleRight,
 ]
