+++ src/app/modules/options/components/options/options.component.html
@@ -58,9 +58,7 @@
       [(ngModel)]="stylesheet"
     >
       <mat-radio-button class="example-radio-button" value="primaryTheme">Style 1</mat-radio-button>
-      <mat-radio-button class="example-radio-button" value="secondaryTheme"
-        >Style 2</mat-radio-button
-      >
+      <mat-radio-button class="example-radio-button" value="codexTheme">Style 2</mat-radio-button>
     </mat-radio-group>
   </div>
   <div fxLayout="column" fxLayoutAlign="space-around" class="maingroup">
