+++ src/app/modules/querybuilder/components/querybuilder-editor/display/display-crit-group/display-crit-group.component.scss
@@ -1,7 +1,7 @@
 .container-crit-group {
   height: 100%;
   padding: 1em 2% 1em 0;
-  border-left: 0.5em solid var(--num-color--accent);
+  border-left: 0.5em solid var(--num-color--accent-600);
 }
 
 .outer {
