+++ src/app/modules/querybuilder/components/querybuilder-editor/display/display-criterion/display-criterion.component.scss
@@ -14,7 +14,7 @@
 
 .code-part {
   color: var(--num-color--accent-contrast);
-  background-color: var(--num-color--accent-200);
+  background-color: var(--num-color--accent-100);
 
   padding: 0.5em;
 }
