+++ src/app/modules/querybuilder/components/querybuilder-editor/display/display-group/display-group.component.scss
@@ -7,7 +7,7 @@
 }
 
 .query-group mat-card-header {
-  background: var(--num-color--accent);
+  background: var(--num-color--accent-600);
   color: var(--num-color--accent-contrast);
   border-top-right-radius: 5px;
   border-top-left-radius: 5px;
