+++ src/app/modules/querybuilder/components/querybuilder-editor/querybuilder-editor.component.scss
@@ -22,7 +22,7 @@
   top: 0;
   height: 4em;
   background-color: white;
-  border: 1px solid black;
+  border: 1px solid var(--num-color--primary);
   border-radius: 5px;
   z-index: 100;
 }
@@ -31,7 +31,7 @@
   position: sticky;
   bottom: 0;
   background-color: white;
-  border: 1px solid black;
+  border: 1px solid var(--num-color--primary);
   border-radius: 5px;
   z-index: 100;
 }
