+++ src/app/modules/querybuilder/components/querybuilder-editor/search/search-input/search-input.component.scss
@@ -1,7 +1,7 @@
 .container-search-input {
   position: relative;
   min-height: 6em;
-  border: 1px solid var(--num-color--accent);
+  border: 1px solid var(--num-color--primary);
   border-radius: 5px;
 }
 
@@ -16,16 +16,22 @@
 .header-label {
   width: 33%;
   min-width: min-content;
-  background-color: var(--num-color--accent-300);
+  background-color: var(--num-color--accent-600);
   color: var(--num-color--accent-contrast);
   text-align: center;
   border-bottom-right-radius: 5px;
 }
 
 .button-tree-view {
-  background-color: var(--num-color--accent-300);
+  background-color: var(--num-color--primary-900);
+}
+.button-tree-view:hover {
+  background-color: var(--num-color--primary);
 }
 
+.ng-fa-icon {
+  color: var(--num-color--primary-100);
+}
 .mat-icon-button {
   height: 44px;
   width: 44px;
