+++ src/app/modules/querybuilder/components/querybuilder-editor/search/search-text-overlay-content/search-text-overlay-content.component.scss
@@ -1,12 +1,12 @@
 .container {
   width: 70vw;
-  min-height: 10em;
-
+  min-height: 12em;
   background-color: white;
+  max-width: 1200px;
 }
 
 .search-results {
-  max-height: 40vh;
+  max-height: 50vh;
   overflow-y: auto;
   padding: 1em;
 }
