+++ src/app/modules/querybuilder/components/querybuilder-editor/search/search-tree-overlay-content/search-tree-overlay-content.component.scss
@@ -1,13 +1,12 @@
 .container {
   width: 70vw;
-  min-height: 10em;
-
+  min-height: 12em;
   background-color: white;
+  max-width: 1200px;
 }
 
 .ontology-tree {
-  max-height: 40vh;
-
+  max-height: 50vh;
   overflow-y: auto;
 }
