+++ src/app/modules/querybuilder/service/backend.service.spec.ts
@@ -19,18 +19,22 @@ describe('BackendService', () => {
   const EXAMPLE_SEARCH = 'DIAB'
   const EXAMPLE_URL = 'http:/abc/querybuillder/result?id=123456'
 
+  const featureService = {
+    getPatientResultLowerBoundary(): number {
+      return 10
+    },
+    mockTerminology(): boolean {
+      return true
+    },
+    mockQuery(): boolean {
+      return true
+    },
+    mockResult(): boolean {
+      return true
+    },
+  } as FeatureService
+
   beforeEach(async () => {
-    const featureService = {
-      mockTerminology(): boolean {
-        return true
-      },
-      mockQuery(): boolean {
-        return true
-      },
-      mockResult(): boolean {
-        return true
-      },
-    } as FeatureService
     await TestBed.configureTestingModule({
       imports: [HttpClientTestingModule],
       providers: [
@@ -48,7 +52,7 @@ describe('BackendService', () => {
   })
 
   it('should return programmatically mocked categories (root entries)', (done: DoneCallback) => {
-    const featureService = TestBed.inject<FeatureService>(FeatureService)
+    // const featureService = TestBed.inject<FeatureService>(FeatureService)
     jest.spyOn(featureService, 'mockTerminology').mockReturnValue(true)
 
     service.getCategories().subscribe((categories: Array<CategoryEntry>) => {
@@ -62,7 +66,7 @@ describe('BackendService', () => {
     jest
       .spyOn(appConfigService, 'getConfig')
       .mockReturnValue(BackendServiceSpecUtil.createConfig('http:/abc'))
-    const featureService = TestBed.inject<FeatureService>(FeatureService)
+    // const featureService = TestBed.inject<FeatureService>(FeatureService)
     jest.spyOn(featureService, 'mockTerminology').mockReturnValue(false)
 
     const httpMock: HttpTestingController = TestBed.inject(HttpTestingController)
@@ -77,7 +81,7 @@ describe('BackendService', () => {
   })
 
   it('should return programmatically mocked terminology tree', (done: DoneCallback) => {
-    const featureService = TestBed.inject<FeatureService>(FeatureService)
+    // const featureService = TestBed.inject<FeatureService>(FeatureService)
     jest.spyOn(featureService, 'mockTerminology').mockReturnValue(true)
 
     service.getTerminolgyTree(EXAMPLE_ID).subscribe((entry: TerminologyEntry) => {
@@ -91,7 +95,7 @@ describe('BackendService', () => {
     jest
       .spyOn(appConfigService, 'getConfig')
       .mockReturnValue(BackendServiceSpecUtil.createConfig('http:/abc'))
-    const featureService = TestBed.inject<FeatureService>(FeatureService)
+    // const featureService = TestBed.inject<FeatureService>(FeatureService)
     jest.spyOn(featureService, 'mockTerminology').mockReturnValue(false)
 
     const httpMock: HttpTestingController = TestBed.inject(HttpTestingController)
@@ -106,7 +110,7 @@ describe('BackendService', () => {
   })
 
   it('should return programmatically mocked search result list', (done: DoneCallback) => {
-    const featureService = TestBed.inject<FeatureService>(FeatureService)
+    // const featureService = TestBed.inject<FeatureService>(FeatureService)
     jest.spyOn(featureService, 'mockTerminology').mockReturnValue(true)
 
     service
@@ -122,7 +126,7 @@ describe('BackendService', () => {
     jest
       .spyOn(appConfigService, 'getConfig')
       .mockReturnValue(BackendServiceSpecUtil.createConfig('http:/abc'))
-    const featureService = TestBed.inject<FeatureService>(FeatureService)
+    // const featureService = TestBed.inject<FeatureService>(FeatureService)
     jest.spyOn(featureService, 'mockTerminology').mockReturnValue(false)
 
     const httpMock: HttpTestingController = TestBed.inject(HttpTestingController)
@@ -147,7 +151,7 @@ describe('BackendService', () => {
     jest
       .spyOn(appConfigService, 'getConfig')
       .mockReturnValue(BackendServiceSpecUtil.createConfig('http:/abc'))
-    const featureService = TestBed.inject<FeatureService>(FeatureService)
+    // const featureService = TestBed.inject<FeatureService>(FeatureService)
     jest.spyOn(featureService, 'mockTerminology').mockReturnValue(false)
 
     const httpMock: HttpTestingController = TestBed.inject(HttpTestingController)
@@ -166,7 +170,7 @@ describe('BackendService', () => {
   })
 
   it('should post programmatically mocked query', (done: DoneCallback) => {
-    const featureService = TestBed.inject<FeatureService>(FeatureService)
+    // const featureService = TestBed.inject<FeatureService>(FeatureService)
     jest.spyOn(featureService, 'mockQuery').mockReturnValue(true)
 
     service.postQuery(new Query()).subscribe((queryResponse: QueryResponse) => {
@@ -180,7 +184,7 @@ describe('BackendService', () => {
     jest
       .spyOn(appConfigService, 'getConfig')
       .mockReturnValue(BackendServiceSpecUtil.createConfig('http:/abc'))
-    const featureService = TestBed.inject<FeatureService>(FeatureService)
+    // const featureService = TestBed.inject<FeatureService>(FeatureService)
     jest.spyOn(featureService, 'mockQuery').mockReturnValue(false)
 
     const httpMock: HttpTestingController = TestBed.inject(HttpTestingController)
@@ -195,7 +199,7 @@ describe('BackendService', () => {
   })
 
   it('should return programmatically mocked result', (done: DoneCallback) => {
-    const featureService = TestBed.inject<FeatureService>(FeatureService)
+    // const featureService = TestBed.inject<FeatureService>(FeatureService)
     jest.spyOn(featureService, 'mockResult').mockReturnValue(true)
 
     service.getResult(EXAMPLE_URL).subscribe((result: QueryResult) => {
@@ -205,7 +209,7 @@ describe('BackendService', () => {
   })
 
   it('should return mocked result', (done: DoneCallback) => {
-    const featureService = TestBed.inject<FeatureService>(FeatureService)
+    // const featureService = TestBed.inject<FeatureService>(FeatureService)
     jest.spyOn(featureService, 'mockResult').mockReturnValue(false)
     const httpMock: HttpTestingController = TestBed.inject(HttpTestingController)
     const mockResponse: QueryResult = {
