+++ src/assets/config/config.deploy.json
@@ -31,13 +31,14 @@
   },
   "options": {
     "pollingtimeinseconds": 10,
-    "pollingintervallinseconds": 1
+    "pollingintervallinseconds": 1,
+    "lowerboundarypatientresult": 10
   },
   "mock": {
     "terminology": false,
     "query": false,
     "result": false
   },
-  "stylesheet": "primaryTheme",
+  "stylesheet": "codexTheme",
   "fhirport": "8082"
 }
