+++ src/styles/custom-material/styles/_color-variables.themed.scss
@@ -4,6 +4,8 @@
   $warn: map-get($theme, warn);
 
   --num-color--primary: #{mat-color($primary)};
+  --num-color--primary-100: #{mat-color($primary, 100)};
+  --num-color--primary-900: #{mat-color($primary, 900)};
   --num-color--accent: #{mat-color($accent, 500)};
   --num-color--warn: #{mat-color($warn)};
 
@@ -11,6 +13,9 @@
   --num-color--accent-100: #{mat-color($accent, 100)};
   --num-color--accent-200: #{mat-color($accent, 200)};
   --num-color--accent-300: #{mat-color($accent, 300)};
+  --num-color--accent-400: #{mat-color($accent, 400)};
+  --num-color--accent-600: #{mat-color($accent, 600)};
+  --num-color--accent-700: #{mat-color($accent, 700)};
 
   --num-color--primary-contrast: #{mat-contrast($primary, default)};
   --num-color--accent-contrast: #{mat-contrast($accent, 500)};
