+++ src/styles/custom-material/styles/_divider-classes.themed.scss
@@ -26,7 +26,7 @@
     &.num-divider-color {
       &--primary,
       &--p {
-        border-color: mat-color($primary);
+        border-color: mat-color($primary, 400);
       }
       &--accent,
       &--a {
