+++ src/styles/custom-material/styles/_footer.themed.scss
@@ -1,8 +1,20 @@
 @mixin footer_themed($theme) {
   $primary: map-get($theme, primary);
+  $accent: map-get($theme, accent);
+  $name: map-get($theme, name);
+  $background: map-get($theme, additional);
 
-  .num-footer {
-    background: mat-color($primary, 900);
-    color: mat-contrast($primary, 900);
+  @if ($name == 'primary') {
+    .num-footer {
+      background: mat-color($accent, 500);
+      color: mat-contrast($primary, 900);
+    }
+  }
+  @if ($name == 'codex') {
+    .num-footer {
+      background: mat-color($background, white);
+      color: mat-contrast($primary, 200);
+      border-top: 1px rgba(0, 0, 0, 0.2) solid;
+    }
   }
 }
