+++ src/styles/custom-material/styles/_side-menu.component.themed.scss
@@ -1,14 +1,35 @@
 @mixin side-menu-component_themed($theme) {
   $accent: map-get($theme, accent);
+  $primary: map-get($theme, primary);
+  $name: map-get($theme, name);
 
-  .mat-nav-list {
-    background: mat-color($accent, 500) !important;
-    color: mat-contrast($accent, 500) !important;
+  @if ($name == 'primary') {
+    .mat-nav-list {
+      background: mat-color($accent, 500) !important;
+      color: mat-contrast($accent, 500) !important;
+    }
+    .num-logo-row a {
+      /*background: mat-color($accent, 500) !important;*/
+      color: mat-contrast($accent, 500) !important;
+    }
+    .num-mat-list-item {
+      /*background: mat-color($accent, 500) !important;*/
+      color: mat-contrast($accent, 500) !important;
+    }
   }
-
-  .num-logo-row a {
-    background: mat-color($accent, 500) !important;
-    color: mat-contrast($accent, 500) !important;
+  @if ($name == 'codex') {
+    .mat-nav-list {
+      background: mat-color($primary, 200) !important;
+      color: mat-contrast($primary, 200) !important;
+    }
+    .num-logo-row a {
+      /*background: mat-color($accent, 500) !important;*/
+      color: mat-contrast($primary, 200) !important;
+    }
+    .num-mat-list-item {
+      /*background: mat-color($accent, 500) !important;*/
+      color: mat-contrast($primary, 200) !important;
+    }
   }
 
   .num-logo-row button:hover {
@@ -17,9 +38,6 @@
   }
 
   .num-mat-list-item {
-    background: mat-color($accent, 500) !important;
-    color: mat-contrast($accent, 500) !important;
-
     font-size: 16px !important;
     font-weight: 400 !important;
     line-height: 20px !important;
