+++ src/styles/styles.scss
@@ -10,7 +10,7 @@
   @include angular-material-theme($num-portal-webapp-theme);
   @include custom-material_themed($num-portal-webapp-theme);
 }
-.secondaryTheme {
+.codexTheme {
   @include angular-material-theme($num-portal-webapp-theme2);
   @include custom-material_themed($num-portal-webapp-theme2);
 }
