+++ src/styles/themes/colors/_codex-accent-theme.palette.scss
@@ -1,18 +1,18 @@
-$num-secondary-palette: (
-  50: #ffebee,
-  100: #ffcdd2,
-  200: #ef9a9a,
-  300: #e57373,
-  400: #ef5350,
-  500: #f44336,
-  600: #e53935,
-  700: #d32f2f,
-  800: #c62828,
-  900: #b71c1c,
-  A100: #ff8a80,
-  A200: #ff5252,
-  A400: #ff1744,
-  A700: #d50000,
+$num-codex-accent-palette: (
+  50: #dbecf8,
+  100: #9cc9eb,
+  200: #6db0e1,
+  300: #3290d5,
+  400: #2780c1,
+  500: #226fa8,
+  600: #1d5e8f,
+  700: #184d75,
+  800: #133d5c,
+  900: #0d2c42,
+  A100: #cbe8fe,
+  A200: #67bdfc,
+  A400: #188ee5,
+  A700: #217ec3,
   contrast: (
     50: #000000,
     100: #000000,
