+++ src/styles/themes/colors/_codex-primary-theme.palette.scss
@@ -1,0 +1,32 @@
+$num-codex-primary-palette: (
+  50: #ffffff,
+  100: #ffffff,
+  200: #e8ebed,
+  300: #c0c7ce,
+  400: #afb8c0,
+  500: #9ea9b3,
+  600: #8d9aa6,
+  700: #7c8b98,
+  800: #6c7b89,
+  900: #5e6c78,
+  A100: #ffffff,
+  A200: #f1f5f9,
+  A400: #b1c3d3,
+  A700: #a9b6c2,
+  contrast: (
+    50: #000000,
+    100: #000000,
+    200: #000000,
+    300: #000000,
+    400: #000000,
+    500: #000000,
+    600: #000000,
+    700: #000000,
+    800: #ffffff,
+    900: #ffffff,
+    A100: #000000,
+    A200: #000000,
+    A400: #000000,
+    A700: #000000,
+  ),
+);
